@@ -8,33 +8,35 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": "^7.2.5|^8.0",
-        "laravel/framework": "^7.29",
+        "php": "^8.0",
+        "laravel/framework": "^8.46",
         "intervention/image": "^2.5",
-        "emojione/emojione": "^4.5",
+        "joypixels/emoji-toolkit": "^6.5",
         "nesk/puphpeteer": "^1.6",
         "sentry/sentry-laravel": "^2.6",
         "fideloper/proxy": "^4.4",
         "laravel/tinker": "^2.5",
         "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^6.3.1|^7.0.1"
+        "guzzlehttp/guzzle": "^7.0.1"
     },
     "require-dev": {
-        "mockery/mockery": "^1.3.1",
-        "phpunit/phpunit": "^8.5.8|^9.3.3",
+        "mockery/mockery": "^1.4.2",
+        "phpunit/phpunit": "^9.3.3",
         "symfony/css-selector": "^5.2",
         "symfony/dom-crawler": "^5.3",
         "friendsofphp/php-cs-fixer": "^3.0",
         "fakerphp/faker": "^1.9.1",
-        "facade/ignition": "^2.0",
-        "nunomaduro/collision": "^4.3"
+        "facade/ignition": "^2.5",
+        "nunomaduro/collision": "^5.0"
     },
     "autoload": {
         "classmap": [
             "database"
         ],
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
         }
     },
     "autoload-dev": {
